cancel created contracts of failed node group
